chore(Housekeeping): Remove axios dependency and dead code
This removes axios since it was fully replaced by fetch. The only
trace we had left in there was in the function fetchAPI which was called
in a dead component called SidebarAdopted. This was not used at all.
Also removes IconRemove which also was used by SidebarAdopted and
noone else.

There where also another dead route called selected.

Also removes it from storybook and updates test snapshots.