docs: demonstrate that new session unmounts PasswordEditForm
When `signInWithPassword` is successful, a new session is created via the `SessionContextProvider`. This causes the `PasswordEditForm` component to unmount, resulting in the `setNotification` to not have an effect anymore: after all the component including the notification UI is unmounted. Note that we still see the console logs because the `updatePassword` function actually finishes inside the `formData.password !== formData.repeatPassword` condition, so we've got our logs with a null notification.